Move reused util funcs to _util and types to constants
