+++ src/ptwt/constants.py
@@ -1,7 +1,9 @@
 """Constants and types used throughout the PyTorch Wavelet Toolbox."""
 
+from __future__ import annotations
+
 from collections.abc import Sequence
-from typing import Literal, NamedTuple, Union
+from typing import Literal, NamedTuple, Protocol, Union
 
 import torch
 from typing_extensions import TypeAlias, Unpack
@@ -17,6 +19,62 @@ __all__ = [
     "WaveletCoeffNd",
     "WaveletDetailDict",
 ]
+
+
+class Wavelet(Protocol):
+    """Wavelet object interface, based on the pywt wavelet object."""
+
+    name: str
+    dec_lo: Sequence[float]
+    dec_hi: Sequence[float]
+    rec_lo: Sequence[float]
+    rec_hi: Sequence[float]
+    dec_len: int
+    rec_len: int
+    filter_bank: tuple[
+        Sequence[float], Sequence[float], Sequence[float], Sequence[float]
+    ]
+
+    def __len__(self) -> int:
+        """Return the number of filter coefficients."""
+        return len(self.dec_lo)
+
+
+class WaveletTensorTuple(NamedTuple):
+    """Named tuple containing the wavelet filter bank to use in JIT code."""
+
+    dec_lo: torch.Tensor
+    dec_hi: torch.Tensor
+    rec_lo: torch.Tensor
+    rec_hi: torch.Tensor
+
+    @property
+    def dec_len(self) -> int:
+        """Length of decomposition filters."""
+        return len(self.dec_lo)
+
+    @property
+    def rec_len(self) -> int:
+        """Length of reconstruction filters."""
+        return len(self.rec_lo)
+
+    @property
+    def filter_bank(
+        self,
+    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
+        """Filter bank of the wavelet."""
+        return self
+
+    @classmethod
+    def from_wavelet(cls, wavelet: Wavelet, dtype: torch.dtype) -> WaveletTensorTuple:
+        """Construct Wavelet named tuple from wavelet protocol member."""
+        return cls(
+            torch.tensor(wavelet.dec_lo, dtype=dtype),
+            torch.tensor(wavelet.dec_hi, dtype=dtype),
+            torch.tensor(wavelet.rec_lo, dtype=dtype),
+            torch.tensor(wavelet.rec_hi, dtype=dtype),
+        )
+
 
 BoundaryMode = Literal["constant", "zero", "reflect", "periodic", "symmetric"]
 """
@@ -36,6 +94,7 @@ This is a type literal for the way of handling signal boundaries.
 This is either a form of padding (see :data:`ptwt.constants.BoundaryMode`
 for padding options) or ``boundary`` to use boundary wavelets.
 """
+
 
 # TODO: Add documentation on the different values of PaddingMode
