+++ src/ptwt/conv_transform_2.py
@@ -13,27 +13,24 @@ import pywt
 import torch
 
 from ._util import (
-    Wavelet,
+    _adjust_padding_at_reconstruction,
     _as_wavelet,
     _check_axes_argument,
     _check_if_tensor,
     _fold_axes,
+    _get_filter_tensors,
     _get_len,
+    _get_pad,
     _is_dtype_supported,
     _map_result,
     _outer,
     _pad_symmetric,
     _swap_axes,
+    _translate_boundary_strings,
     _undo_swap_axes,
     _unfold_axes,
 )
-from .constants import BoundaryMode, WaveletCoeff2d, WaveletDetailTuple2d
-from .conv_transform import (
-    _adjust_padding_at_reconstruction,
-    _get_filter_tensors,
-    _get_pad,
-    _translate_boundary_strings,
-)
+from .constants import BoundaryMode, Wavelet, WaveletCoeff2d, WaveletDetailTuple2d
 
 
 def _construct_2d_filt(lo: torch.Tensor, hi: torch.Tensor) -> torch.Tensor:
