+++ src/ptwt/separable_conv_transform.py
@@ -16,7 +16,6 @@ import numpy as np
 import torch
 
 from ._util import (
-    Wavelet,
     _as_wavelet,
     _check_axes_argument,
     _check_if_tensor,
@@ -27,7 +26,7 @@ from ._util import (
     _undo_swap_axes,
     _unfold_axes,
 )
-from .constants import BoundaryMode, WaveletCoeff2dSeparable, WaveletCoeffNd
+from .constants import BoundaryMode, Wavelet, WaveletCoeff2dSeparable, WaveletCoeffNd
 from .conv_transform import wavedec, waverec
 from .conv_transform_2 import _preprocess_tensor_dec2d
