+++ tests/test_convolution_fwt.py
@@ -10,14 +10,9 @@ import pywt
 import torch
 from scipy import datasets
 
-from ptwt._util import _outer
+from ptwt._util import _flatten_2d_coeff_lst, _outer
 from ptwt.constants import BoundaryMode
-from ptwt.conv_transform import (
-    _flatten_2d_coeff_lst,
-    _translate_boundary_strings,
-    wavedec,
-    waverec,
-)
+from ptwt.conv_transform import _translate_boundary_strings, wavedec, waverec
 from ptwt.conv_transform_2 import wavedec2, waverec2
 from ptwt.wavelets_learnable import SoftOrthogonalWavelet
 from tests._mackey_glass import MackeyGenerator
